+++ packages/terra-clinical-result/CHANGELOG.md
@@ -2,9 +2,16 @@
 
 ## Unreleased
 
+* Added
+  * Added `width: 100%` to flowsheet result cell styles
+
+* Changed
+  * Updated useLayoutEffect to useEffect
+  * Removed unnecessary dependencies and component states from useEffect
+
 ## 1.21.0 - (August 2, 2023)
 
-* Changed  
+* Changed
   * Locked `uuid` dependency to `3.4.0` for maximum consistency across Terra packages.
   * Updated translations.