+++ packages/terra-clinical-result/CHANGELOG.md
@@ -5,9 +5,8 @@
 * Added
   * Added back `width: 100%` to flowsheet result cell styles removed in 1.17.0.
 
-* Changed
-  * Updated FlowsheetResultCell useLayoutEffect to useEffect.
-  * Updated FlowsheetResultCell  dependencies and component states from useEffect.
+* Fixed:
+  * Fixed infinite re-render bug.
 
 ## 1.21.0 - (August 2, 2023)